+++ src/main/java/com/javarush/Main.java
@@ -8,17 +8,18 @@ import java.util.Arrays;
 public class Main {
     public static void main(String[] args) {
         Runner runner = new Runner();
+        CLI cli = new CLI();
         if (args.length == 2 || args.length == 3) {
             String command = args[0];
             String filePath = args[1];
             int key = Integer.parseInt(args[3]);
             runner.run();
         } else {
-            CLI cli = new CLI();
             cli.InputUser();
             runner.run();
 
         }
+
         System.out.println(Arrays.toString(args));
     }
 }
