+++ src/main/java/com/javarush/cli/CLI.java
@@ -3,16 +3,22 @@ package com.javarush.cli;
 import java.util.Scanner;
 
 public class CLI {
+    public   String command;
+    public String filePath;
+    public int key;
 
-    public void InputUser(){
+    public void InputUser() {
 
         System.out.println("Any arguments were entered from keyboard...");
         System.out.println("Please enter a command [ENCRYPTED]/[DECRYPTED]/[BRUTE_FORCE]:");
         Scanner scanner = new Scanner(System.in);
-        String command = scanner.nextLine();
+        this.command = scanner.nextLine();
         System.out.println("Please enter a source/dest filepath: ");
-        String filePath = scanner.nextLine();
+        filePath = scanner.nextLine();
         System.out.println("Please enter a key:");
-        int key = scanner.nextInt();
+        key = scanner.nextInt();
+        System.out.println(command);
+        System.out.println(filePath);
+        System.out.println(key);
     }
 }
