+++ src/main/java/com/javarush/io/FileReader.java
@@ -3,15 +3,12 @@ package com.javarush.io;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 
-public class FileService {
+public class FileReader {
     private String filePath;
 
 
-    public FileService(String filePath) {
+    public FileReader(String filePath) {
         this.filePath = filePath;
     }
 
@@ -29,16 +26,6 @@ public class FileService {
             System.out.println("Oops! " + e.getMessage());
         }
         return readString.toString();
-    }
-
-    public void writeFile(String encryptedText) {
-        try (OutputStream fos = Files.newOutputStream(Paths.get(filePath + " [ENCRYPTED]"))) {
-            byte[] bytes = encryptedText.getBytes();
-            fos.write(bytes);
-        } catch (IOException ex) {
-            System.out.println("Oops! " + ex.getMessage());
-
-        }
     }
 }
