+++ src/main/java/com/javarush/runner/Runner.java
@@ -1,9 +1,13 @@
 package com.javarush.runner;
 
 public class Runner {
-    public void run(){
+
+
+    public void run() {
+
         System.out.println("Runner is working...");
+        System.out.println();
+
+
     }
-
-
 }
